% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{adj2ig}
\alias{adj2ig}
\title{Create an \code{igraph} object from weighted adjacency matrix}
\usage{
adj2ig(w.adj, v.col = NULL, e.col = NULL)
}
\arguments{
\item{w.adj}{The weighted adjacency matrix to be converted to an \code{igraph} object}

\item{v.col}{The vertex colours. Numeric vector of length equal to number of taxa in network.}

\item{e.col}{The edge colours.  Numeric vector of length equal to number of taxa in network. If NULL, then colour green/red based on positive/negative weights.}
}
\value{
An object of class \code{igraph} corresponding to the provided adjacency matrix
}
\description{
Create an \code{igraph} object from weighted adjacency matrix
}
\examples{

\donttest{
library(mdine)
data(crohns)

X <- model.matrix(~disease, data=crohns$covars)
md.fit <- mdine(Y=crohns$otu.counts, X=X, Z=X[,2], mc.cores=1, iter=100)
adj <- ci2adj(md.fit, weighted = TRUE)

ig0 <- adj2ig(adj$adj0)
igraph::plot.igraph(ig0)
}

}
