% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{ci2adj}
\alias{ci2adj}
\title{Get adjacency matrix for estimated networks based on credible intervals}
\usage{
ci2adj(obj, weighted = FALSE)
}
\arguments{
\item{obj}{An object of class \code{mdine}}

\item{weighted}{Logical: should the adjacency matrices be weighted according to the corresponding partial correlation matrix?}
}
\value{
A list containing adjacency matrices for the two estimated precision matrices
}
\description{
Get adjacency matrix for estimated networks based on credible intervals
}
\examples{
ls()
}
