% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crohns-data.R
\docType{data}
\name{crohns}
\alias{crohns}
\title{Microbiome family-level data from Crohn's and control samples}
\format{List with two objects: \code{covars} is a dataframe with 100 rows and 4 columns containing the covariate data; \code{otu.counts} is a matrix with 100 rows and 6 columns which contains the family-level counts of each subject.  The last column of \code{otu.counts} is a reference category containing the sum of all families not included in this dataset.}
\source{
\href{https://qiita.ucsd.edu/study/description/1939}{Qiita database}
}
\usage{
data(crohns)
}
\description{
Counts from 16S sequencing data on treatment-naive subjects with Crohn's disease
or controls.  Other covariates available include age, sex, and antibiotic use.
}
\examples{
data(crohns)

}
\references{
Gevers et al. (2014) Cell Host & Microbe 15.3 (2014): 382-392.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/24629344}{PubMed})
}
\keyword{crohn's}
\keyword{datasets}
\keyword{microbiome}
