% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdine-package.R
\docType{package}
\name{mdine-package}
\alias{mdine-package}
\alias{mdine}
\title{The 'mdine' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
