% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{plot_networks}
\alias{plot_networks}
\title{Plot a network for each group}
\usage{
plot_networks(obj)
}
\arguments{
\item{obj}{An object of class \code{mdine}}
}
\value{
No return value
}
\description{
Plot a network for each group
}
\examples{
ls()
}
