% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{adj2ig}
\alias{adj2ig}
\title{Create an \code{igraph} object from weighted adjacency matrix}
\usage{
adj2ig(w.adj = NULL, v.col = NULL, e.col = NULL)
}
\arguments{
\item{w.adj}{The weighted adjacency matrix to be converted to an \code{igraph} object}

\item{v.col}{The vertex colours.  Numeric vector of length equal to number of OTUs in network}

\item{e.col}{The edge colour.  If NULL, then colour green/red based on positive/negative weights.}
}
\value{
An object of class \code{igraph} corresponding to the provided adjacency matrix
}
\description{
Create an \code{igraph} object from weighted adjacency matrix
}
\examples{
 ls()
}
