% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdine-package.R
\docType{package}
\name{mdine-package}
\alias{mdine-package}
\title{The 'mdine' package.}
\description{
Estimates precision matrix networks for two groups (as well as their difference) in microbiome sequencing data.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. \href{http://mc-stan.org}{http://mc-stan.org}
}
