% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdine.R
\name{mdine}
\alias{mdine}
\title{Microbiome differential network estimation}
\usage{
mdine(Y, X, Z, lambda = NULL, offset = NULL, mc.cores = 1,
  iter = 1000, chains = 4, quant = c(0.025, 0.975), ...)
}
\arguments{
\item{Y}{The (unrarefied) taxa count matrix with rows as samples and columns as taxa.  The last column is
the reference category, and is not included in the estimated network.}

\item{X}{The model matrix (including an intercept column)}

\item{Z}{Binary variable over which the network is assumed to vary.}

\item{lambda}{Network penalization parameter.  If NULL, then lambda is estimated}

\item{offset}{A vector containing an offset term for each subject}

\item{mc.cores}{The number of cores to run MCMC chains in parallel}

\item{iter}{The number of MCMC iterations.  By default the first half of the iterations will be used as warmup.}

\item{chains}{The number of MCMC chains.}

\item{quant}{Lower and upper quantiles of the posterior distribution to create credible intervals.}

\item{...}{Other arguments passed to `rstan::sampling`}
}
\value{
A an object of class "mdine" containing posterior means for the model parameters, credible intervals,
and the stanfit object.
}
\description{
Microbiome differential network estimation
}
\examples{
ls()
}
