% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn_model.R
\name{mn_model}
\alias{mn_model}
\title{TEST}
\usage{
mn_model(x, y, ...)
}
\arguments{
\item{x}{Numeric vector of input values.}

\item{y}{Numberic vector of output values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
TEST
}
