% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{sig_diff_prec}
\alias{sig_diff_prec}
\title{Get which elements of the precision matrices differ significantly}
\usage{
sig_diff_prec(obj)
}
\arguments{
\item{obj}{An object of class \code{mdine}}
}
\value{
A matrix with a 1 if that element of the precision matrix differs significantly with respect to the binary covariate
}
\description{
Get which elements of the precision matrices differ significantly
}
\examples{
ls()
}
